Create + delete test
