+++ src/main/resources/application.properties
@@ -3,6 +3,6 @@ spring.datasource.url=jdbc:mysql://localhost:3306/kijani
 spring.datasource.username=root
 spring.datasource.password=12345678
 spring.jpa.properties.hibernate.jdbc.time_zone=UTC
-spring.jpa.hibernate.ddl-auto=create-drop
+spring.jpa.hibernate.ddl-auto=update
 spring.jpa.open-in-view=false
 spring.data.jpa.repositories.bootstrap_mode=default
